+++ spec/transpec/syntax/matcher_spec.rb
@@ -131,6 +131,44 @@ module Transpec
           end
         end
 
+        context 'when it is `==1` form' do
+          let(:source) do
+            <<-END
+              it 'is 1' do
+                subject.should==1
+              end
+            END
+          end
+
+          let(:expected_source) do
+            <<-END
+              it 'is 1' do
+                subject.should eq(1)
+              end
+            END
+          end
+
+          it 'converts into `eq(1)` form' do
+            rewritten_source.should == expected_source
+          end
+
+          context 'and false is passed as `parenthesize_arg` argument' do
+            let(:parenthesize_arg) { false }
+
+            let(:expected_source) do
+            <<-END
+              it 'is 1' do
+                subject.should eq 1
+              end
+            END
+            end
+
+            it 'converts into `eq 1` form' do
+              rewritten_source.should == expected_source
+            end
+          end
+        end
+
         context 'when it is `be == 1` form' do
           let(:source) do
             <<-END
@@ -315,6 +353,28 @@ module Transpec
           end
         end
 
+        context 'when it is `=~/pattern/` form' do
+          let(:source) do
+            <<-END
+              it 'matches the pattern' do
+                subject.should=~/pattern/
+              end
+            END
+          end
+
+          let(:expected_source) do
+            <<-END
+              it 'matches the pattern' do
+                subject.should match(/pattern/)
+              end
+            END
+          end
+
+          it 'converts into `match(/pattern/)` form' do
+            rewritten_source.should == expected_source
+          end
+        end
+
         context 'when it is `be =~ /pattern/` form' do
           let(:source) do
             <<-END
@@ -342,6 +402,28 @@ module Transpec
             <<-END
               it 'contains 1 and 2' do
                 subject.should =~ [1, 2]
+              end
+            END
+          end
+
+          let(:expected_source) do
+            <<-END
+              it 'contains 1 and 2' do
+                subject.should match_array([1, 2])
+              end
+            END
+          end
+
+          it 'converts into `match_array([1, 2])` form' do
+            rewritten_source.should == expected_source
+          end
+        end
+
+        context 'when it is `=~[1, 2]` form' do
+          let(:source) do
+            <<-END
+              it 'contains 1 and 2' do
+                subject.should=~[1, 2]
               end
             END
           end