+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Master
 
 * Avoid confusing `Excon.stub` with RSpec's `stub` ([#4](https://github.com/yujinakayama/transpec/issues/4))
+* Fix a bug where `be == 1` was converted into `be eq(1)` (now it's converted into `eq(1)`)
 
 ## v0.1.2